snapshot/requirements.txt
python-decouple==3.8
requests==2.22.0
web3==6.0.0
